Render tabs in GB.ts init()
